backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-use-email-as-username==1.4.0
djangorestframework==3.15.1
mysqlclient==2.2.4
PyJWT==2.8.0
sqlparse==0.5.0
tzdata==2024.1
